Remove 'bin' and add id_pool + graceful shutdown
Now client and server are part of the library, not bynaries on
their own.

The id_pool is now a Vector from which the processes are able to
retrieve and give their ids.

The graceful shutdown is implemented taking into account that the
io operation of tokio are blocking, and thus that the runtime
itself must be controlled with a `shutdown_timeout`.